Add -Fformat flag to change tar header format
Add a -Fformat (or -F format) flag to mtar to allow changing the tar
header for the output stream. If the header format changes mid-stream,
a warning is issued.

By default, mtar already uses PAX and will continue to use PAX by
default. This flag is added specifically to support older software that
may not handle PAX.